Adds lazy_views module for lazy-loading views displays.
Adds cache plugin for views and a post-render hook to override the
view display output with a placeholder that then lazy-loads the
real view output. The purpose of this is for helping with being
able to cache output that may be different for different users,
even if they're anonymous. Adds spin.js library. Adds settings
for outputting a spinner and/or placeholder text before the
content is loaded.